qa-portal/backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Base de datos
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Autenticación
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.17

# Configuración
pydantic[email]==2.10.3
pydantic-settings==2.6.1

# Cliente IA (compatible con OpenAI)
openai==1.57.4

# HTTP requests (para APIs externas: Zoho, GitHub, SonarQube)
httpx==0.28.1

# Selenium (imagen scraping de QEngine)
selenium==4.27.1
webdriver-manager==4.0.2
Pillow==11.0.0

# Utilidades
python-dotenv==1.0.1
